+++ CHANGELOG.md
@@ -1,6 +1,8 @@
-## [0.0.1] - 2020-10-8
+## [1.0.1] - 02-11-2020
 
-* Blur is a wrapper widget that blur it's child. There are ImageBlur.asset and ImageBlur.network that blur the image.
+* Add: `padding` property to `Frost` widget.
+* Add: `Frost.text` which takes string as input.
+* Add: `Frost.icon` which takes iconData as input.
 
 ## [1.0.0] - 02-11-2020
 
@@ -9,3 +11,7 @@
 * Add: `Frost` class which blurs the background of the child.
 * Breaking: The `overlayChild` is renamed to `overlay`
 * Add: `alignment` property for the overlay child.
+
+## [0.0.1] - 2020-10-8
+
+* Blur is a wrapper widget that blur it's child. There are ImageBlur.asset and ImageBlur.network that blur the image.